refactor dpm pm quality policy routes
