+++ src/app/router_registry.py
@@ -18,6 +18,9 @@ from app.routers.dpm_command_center_pm_quality import (
 from app.routers.dpm_command_center_pm_quality_fairness import (
     router as dpm_command_center_pm_quality_fairness_router,
 )
+from app.routers.dpm_command_center_pm_quality_policies import (
+    router as dpm_command_center_pm_quality_policies_router,
+)
 from app.routers.dpm_command_center_pm_quality_review_actions import (
     router as dpm_command_center_pm_quality_review_actions_router,
 )
@@ -85,6 +88,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_outcome_reviews_router)
     app.include_router(dpm_command_center_pm_quality_router)
     app.include_router(dpm_command_center_pm_quality_fairness_router)
+    app.include_router(dpm_command_center_pm_quality_policies_router)
     app.include_router(dpm_command_center_pm_quality_review_actions_router)
     app.include_router(dpm_command_center_pm_quality_summary_invocations_router)
     app.include_router(dpm_wave_campaign_definitions_router)
